storage整理

+++ Sources/Storage/Storages/FileStorage.swift
@@ -14,83 +14,58 @@ public struct FileStorage {
         }
     }
     
-    private func upsert<T: Codable>(_ url: URL, value: T) throws {
+    func toURL(_ key: StorageKey) throws -> URL {
+        if let folder = key.folder {
+            return try root.add(folder: folder).appendingPathComponent(key.file)
+        }
+
+        return root.appendingPathComponent(key.file)
+    }
+}
+
+
+extension FileStorage: IStorage {
+
+    public func upsert<T: Codable>(_ key: StorageKey, value: T) throws {
+        let url = try toURL(key)
         let data = try JSONEncoder().encode(value)
+
         let isSuccess = fileManager.createFile(atPath: url.path, contents: data)
         if !isSuccess {
             logging(.warning, "ファイル作成に失敗, key: \(url.absoluteString)")
             throw StorageError.createFileFailed(key: url.absoluteString)
         }
     }
-    
-    private func get<T: Codable>(_ url: URL, type: T.Type) throws -> T {
+
+    public func get<T: Codable>(_ key: StorageKey, type: T.Type) throws -> T {
+        let url = try toURL(key)
         guard fileManager.fileExists(atPath: url.path) else { throw StorageError.notFound(key: url.absoluteString) }
         let data = try Data(contentsOf: url)
         return try JSONDecoder().decode(type, from: data)
     }
-    
-    private func delete(_ url: URL) throws {
-        do {
-            try fileManager.removeItem(at: url)
-        } catch CocoaError.fileNoSuchFile {
-            // 削除対象のファイルがなかった場合は例外を握りつぶす
-            return
-        }
-    }
-}
 
-
-extension FileStorage: IStorage {
-    public func upsert<T: Codable>(key: String, value: T) throws {
-        try upsert(root.appendingPathComponent(key), value: value)
-    }
-
-    public func upsert<T: Codable>(folder: String, key: String, value: T) throws {
-        try upsert(try root.add(folder: folder).appendingPathComponent(key), value: value)
-    }
-
-    public func get<T: Codable>(key: String, type: T.Type) throws -> T {
-        try get(root.appendingPathComponent(key), type: type)
-    }
-    
-    public func get<T: Codable>(folder: String, key: String, type: T.Type) throws -> T {
-        try get(try root.add(folder: folder).appendingPathComponent(key), type: type)
-    }
-
-    public func gets<T: Codable>(folder: String, type: T.Type) throws -> [T] {
-        let files = try root.add(folder: folder).getChildren()
-        
-        return files.compactMap {
-            guard let data = try? Data(contentsOf: $0) else { return nil }
-            return try? JSONDecoder().decode(type, from: data)
-        }
+    public func getKeys(folder: String) throws -> [StorageKey] {
+        try root.add(folder: folder)
+                .getChildren()
+                .map { StorageKey(folder: folder, file: $0.lastPathComponent) }
     }
     
     
-    public func delete(key: String) throws {
-        try delete(root.appendingPathComponent(key))
-    }
-    
-    public func delete(folder: String, key: String) throws {
-        try delete(try root.add(folder: folder).appendingPathComponent(key))
-    }
-    
-    public func deletes(folder: String) throws {
-        let files = try root.add(folder: folder).getChildren()
-        try files.forEach { try delete($0) }
-
+    public func delete(_ key: StorageKey) throws {
+        let url = try toURL(key)
+        // 削除対象のファイルがなかった場合は例外扱いにしない
+        guard fileManager.fileExists(atPath: url.path) else { return }
+        try fileManager.removeItem(at: url)
     }
     
     public func deleteAll() throws {
-        let files = try root.getChildren()
-        try files.forEach { try delete($0) }
-    }
-    
-
-}
+        try root.getChildren()
+                .forEach { try fileManager.removeItem(at: $0) }
+    }}
 
 extension URL {
     fileprivate func getChildren() throws -> [URL] {
         try FileManager.default.contentsOfDirectory(at: self, includingPropertiesForKeys: nil, options: [])
     }
 }
+
